Converter Patch Fix
Currently, some of the converter options are commented out in the patches- but only partially.
This caused buggy converter options that could not be turned off, removed from the part, and did not do the conversions they were meant to do on all ISRU converter units.
This simply completely comments out those options, eliminating the bug.